BoP fix to end all fixes (2/2)
In order to lowercase a folder in a path, all files must be deleted in
one commit, then replaced in the next with a different path...
Also, the old BiomesOPlenty path is used in 1.5.2, and the biomesoplenty
path is used in 1.6.2.

Something also happened where the lowercase old path remained, and the
new path would not clone correctly in windows because the folder names
are identical (aside from the caps).

I just deleted all, then reset with a clean texture set.